refactor(video): :recycle: Moves VideoError/RenderError out of Core/Errors.hpp
VideoError and RenderError don't belong to Core itself -- they describe
Video/VideoSystem and Video/Graphics/Renderer failures respectively.
Splits each into its own VideoError.hpp / RenderError.hpp next to the
subsystem it describes, following the <Name>Error.hpp convention;
Core/Errors.hpp keeps only the generic Result<T>/BoolResult/
REGISTER_ASGE_ERROR machinery plus the errors that do belong to Core
(FileWatcherError, ConfError). Also fixes RenderError's fallback
ToErrorString message, which was copy-pasted from VideoError's.
