+++ src/ASGE/Core/Errors.hpp
@@ -310,55 +310,9 @@ inline str::String ToErrorString(ConfError e) noexcept
     return "uknown configuration error";
 }
 
-// ---------------------------------------------------------------------------------------------
-// VIDEO SYSTEM ERRORS
-// ---------------------------------------------------------------------------------------------
-
-enum class VideoError : std::uint8_t
-{
-    SubsystemInitFailed = 1,
-    WindowCreationFailed,
-    RendererCreationFailed
-};
-
-inline str::String ToErrorString(VideoError e) noexcept
-{
-    switch (e)
-    {
-    case VideoError::SubsystemInitFailed: return "failed to initialize the video subsystem";
-    case VideoError::WindowCreationFailed: return "failed to create the window";
-    case VideoError::RendererCreationFailed: return "failed to create the renderer";
-    }
-    return "uknown video error";
-}
-
-// ---------------------------------------------------------------------------------------------
-// RENDER SYSTEM ERRORS
-// ---------------------------------------------------------------------------------------------
-
-enum class RenderError : std::uint8_t
-{
-    RenderRectFailed = 1,
-    RenderLineFailed,
-    RenderCircleFailed
-};
-
-inline str::String ToErrorString(RenderError e) noexcept
-{
-    switch (e)
-    {
-    case RenderError::RenderRectFailed: return "failed to render the rect";
-    case RenderError::RenderLineFailed: return "failed to render the line";
-    case RenderError::RenderCircleFailed: return "failed to render the circle";
-    }
-    return "uknown video error";
-}
-
 }
 
 // REGISTERING ERRORS CATEGORIES TO THE ERROR DB
 
 REGISTER_ASGE_ERROR(asge::errors::FileWatcherError, "asge.filewatcher")
 REGISTER_ASGE_ERROR(asge::errors::ConfError, "asge.configuration")
-REGISTER_ASGE_ERROR(asge::errors::VideoError, "asge.video")
-REGISTER_ASGE_ERROR(asge::errors::RenderError, "asge.render")
